+++ src/components/3d/index.js
@@ -1,3 +1,2 @@
-export { default as CouncilScene } from './CouncilScene';
-export { default as AgentModel } from './AgentModel';
-export { default as CouncilTable } from './CouncilTable';
+export { default as CouncilScene } from './CouncilScene.jsx';
+export { default as CouncilChamber } from './CouncilChamber.jsx';